pyproject.toml
[project]
name = "firstproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "fastmcp>=2.14.0",
    "mcp[cli]>=1.23.3",
    "reportlab>=4.4.6",
]
